FIX: rmc/server.go converting argument of bit shift to uint - older version of the go compiler was complaining about it.

+++ pkg/sync/rmc/server.go
@@ -113,7 +113,7 @@ func (s *server) finishedRMC(u gomel.Unit, _ gomel.Dag) error {
 func hashToInt64(hash gomel.Hash) int64 {
 	var result int64
 	for p, v := range hash[:] {
-		result += int64(v) * (1 << p)
+		result += int64(v) * (1 << uint(p))
 	}
 	return result
 }
